requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
requests
python-dotenv
